feat: Add LlamaEagleSpeculatorConfig for EAGLE1 and HASS models
Add configuration classes for EAGLE1 and HASS speculator models without the model implementation. This separates the config from the actual model code.

- Add LlamaDecoderParameters for Llama decoder configuration
- Add LlamaEagleSpeculatorConfig with support for both EAGLE1 and HASS variants
- Include config-only unit tests
