+++ src/speculators/models/__init__.py
@@ -1,11 +1,10 @@
 from .independent import IndependentSpeculatorConfig
-from .llama_eagle import LlamaEagleSpeculator, LlamaEagleSpeculatorConfig
+from .llama_eagle_config import LlamaEagleSpeculatorConfig
 from .mlp import MLPSpeculatorConfig
 from .transformer import TransformerSpeculatorConfig
 
 __all__ = [
     "IndependentSpeculatorConfig",
-    "LlamaEagleSpeculator",
     "LlamaEagleSpeculatorConfig",
     "MLPSpeculatorConfig",
     "TransformerSpeculatorConfig",
