--- tests/unit/models/test_llama_eagle_model.py
@@ -1,198 +1,0 @@
-"""Unit tests for LlamaEagleSpeculator model."""
-
-import pytest
-import torch
-from torch import nn
-
-from speculators.models.llama_eagle import (
-    LlamaDecoderParameters,
-    LlamaEagleSpeculator,
-    LlamaEagleSpeculatorConfig,
-)
-
-
-class TestLlamaEagleSpeculator:
-    """Test suite for LlamaEagleSpeculator model."""
-
-    @pytest.fixture
-    def eagle_v1_config(self):
-        """Create a test EAGLE v1 configuration."""
-        return LlamaEagleSpeculatorConfig(
-            num_hidden_layers=2,
-            llama_decoder_params=LlamaDecoderParameters(
-                hidden_size=256,
-                intermediate_size=512,
-                num_attention_heads=8,
-                num_key_value_heads=4,
-                vocab_size=1000,
-                max_position_embeddings=512,
-                mlp_bias=False,
-            ),
-        )
-
-    @pytest.fixture
-    def hass_config(self):
-        """Create a test HASS configuration."""
-        return LlamaEagleSpeculatorConfig(
-            num_hidden_layers=2,
-            fc_bias=True,  # HASS uses bias in fusion layer
-            use_extra_layernorms=True,  # HASS uses extra layernorms
-            extra_layernorm_positions=["post_embedding", "pre_lm_head"],
-            llama_decoder_params=LlamaDecoderParameters(
-                hidden_size=256,
-                intermediate_size=512,
-                num_attention_heads=8,
-                num_key_value_heads=4,
-                vocab_size=1000,
-                max_position_embeddings=512,
-                mlp_bias=False,
-            ),
-        )
-
-    def test_model_initialization_eagle_v1(self, eagle_v1_config):
-        """Test EAGLE v1 model initialization."""
-        model = LlamaEagleSpeculator(eagle_v1_config)
-
-        # Check basic attributes
-        assert model.config == eagle_v1_config
-        assert model.vocab_size == 1000
-
-        # Check fusion layer
-        assert isinstance(model.fc, nn.Linear)
-        assert model.fc.in_features == 256 * 2  # hidden_size * 2
-        assert model.fc.out_features == 256
-        assert model.fc.bias is None  # No bias for EAGLE v1
-
-        # Check embeddings
-        assert isinstance(model.embed_tokens, nn.Embedding)
-        assert model.embed_tokens.num_embeddings == 1000
-        assert model.embed_tokens.embedding_dim == 256
-
-        # Check that extra layernorms are not created
-        assert not hasattr(model, "extra_layernorms")
-
-        # Check first layer normalization is Identity
-        assert isinstance(model.layers[0].input_layernorm, nn.Identity)
-
-        # Check decoder layers
-        assert len(model.layers) == 2
-
-    def test_model_initialization_hass(self, hass_config):
-        """Test HASS model initialization."""
-        model = LlamaEagleSpeculator(hass_config)
-
-        # Check fusion layer has bias
-        assert model.fc.bias is not None
-
-        # Check extra layernorms are created
-        assert hasattr(model, "extra_layernorms")
-        assert "post_embedding" in model.extra_layernorms
-        assert "pre_lm_head" in model.extra_layernorms
-
-    def test_forward_pass_shapes(self, eagle_v1_config):
-        """Test forward pass output shapes."""
-        model = LlamaEagleSpeculator(eagle_v1_config)
-        model.eval()
-
-        batch_size = 2
-        seq_length = 10
-
-        # Create dummy inputs
-        input_ids = torch.randint(0, 1000, (batch_size, seq_length))
-        hidden_states = torch.randn(batch_size, seq_length, 256)
-
-        with torch.no_grad():
-            logits = model(input_ids=input_ids, hidden_states=hidden_states)
-
-        # Check output shape
-        assert logits.shape == (batch_size, seq_length, 1000)
-
-    def test_forward_pass_with_attention_mask(self, eagle_v1_config):
-        """Test forward pass with attention mask."""
-        model = LlamaEagleSpeculator(eagle_v1_config)
-        model.eval()
-
-        batch_size = 2
-        seq_length = 10
-
-        # Create inputs
-        input_ids = torch.randint(0, 1000, (batch_size, seq_length))
-        hidden_states = torch.randn(batch_size, seq_length, 256)
-        attention_mask = torch.ones(batch_size, seq_length)
-        attention_mask[0, 5:] = 0  # Mask out last 5 tokens for first sample
-
-        with torch.no_grad():
-            logits = model(
-                input_ids=input_ids,
-                hidden_states=hidden_states,
-                attention_mask=attention_mask
-            )
-
-        assert logits.shape == (batch_size, seq_length, 1000)
-
-    def test_gradient_flow(self, eagle_v1_config):
-        """Test that gradients flow through the model."""
-        model = LlamaEagleSpeculator(eagle_v1_config)
-        model.train()
-
-        # Create inputs
-        input_ids = torch.randint(0, 1000, (1, 5))
-        hidden_states = torch.randn(1, 5, 256, requires_grad=True)
-
-        # Forward pass
-        logits = model(input_ids=input_ids, hidden_states=hidden_states)
-
-        # Create dummy loss
-        loss = logits.sum()
-        loss.backward()
-
-        # Check gradients exist
-        assert hidden_states.grad is not None
-        assert model.fc.weight.grad is not None
-        assert model.embed_tokens.weight.grad is not None
-
-    def test_get_set_input_embeddings(self, eagle_v1_config):
-        """Test get/set input embeddings methods."""
-        model = LlamaEagleSpeculator(eagle_v1_config)
-
-        # Get embeddings
-        embeddings = model.get_input_embeddings()
-        assert embeddings is model.embed_tokens
-
-        # Set new embeddings
-        new_embeddings = nn.Embedding(2000, 256)
-        model.set_input_embeddings(new_embeddings)
-        assert model.get_input_embeddings() is new_embeddings
-
-    def test_pre_lm_head_layernorm(self):
-        """Test that pre_lm_head extra layernorm works as final layer norm."""
-        # Test configuration with pre_lm_head layernorm
-        config = LlamaEagleSpeculatorConfig(
-            num_hidden_layers=1,
-            use_extra_layernorms=True,
-            extra_layernorm_positions=["pre_lm_head"],
-            llama_decoder_params=LlamaDecoderParameters(
-                hidden_size=256,
-                intermediate_size=512,
-                num_attention_heads=8,
-                num_key_value_heads=4,
-                vocab_size=1000,
-            ),
-        )
-        model = LlamaEagleSpeculator(config)
-        
-        # Check that pre_lm_head layernorm is created
-        assert hasattr(model, "extra_layernorms")
-        assert "pre_lm_head" in model.extra_layernorms
-        assert isinstance(model.extra_layernorms["pre_lm_head"], nn.Module)
-        
-        # Test forward pass to ensure it's applied
-        batch_size = 1
-        seq_length = 5
-        input_ids = torch.randint(0, 1000, (batch_size, seq_length))
-        hidden_states = torch.randn(batch_size, seq_length, 256)
-        
-        with torch.no_grad():
-            output = model(input_ids=input_ids, hidden_states=hidden_states)
-        
-        assert output.shape == (batch_size, seq_length, 1000)
